Moved logging logic to Logs module
